requirements.txt
Django
djangorestframework
Pillow
djangorestframework_simplejwt
drf-yasg